BJ-922: asd (#9)
