+++ src/main/java/com/rbkmoney/threeds/server/config/commonplatform/DsRequestHandlerConfig.java
@@ -1,4 +1,4 @@
-package com.rbkmoney.threeds.server.config.testplatform;
+package com.rbkmoney.threeds.server.config.commonplatform;
 
 import com.rbkmoney.threeds.server.converter.MessageToErrorResConverter;
 import com.rbkmoney.threeds.server.domain.root.Message;
@@ -8,14 +8,12 @@ import com.rbkmoney.threeds.server.handle.DsRequestHandler;
 import com.rbkmoney.threeds.server.handle.impl.UnsupportedMessageTypeRequestHandler;
 import com.rbkmoney.threeds.server.processor.Processor;
 import com.rbkmoney.threeds.server.service.MessageValidatorService;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import static com.rbkmoney.threeds.server.config.builder.HandlerBuilder.createDsRequestHandler;
 
-@Configuration
-@ConditionalOnProperty(name = "preparation-flow.mode", havingValue = "TEST_PLATFORM")
+@Configuration("commonDsRequestHandlerConfig")
 public class DsRequestHandlerConfig {
 
     @Bean
